Simplify JnlpProtocolHandlerTest.
Some of this is similar to the simplifications for NetworkLayerTest in e9c1af14. This test has also been hitting sporadic hangs / failures, primarily on certain OSes, etc. These mostly occur in preClose0() also. Similar to that test, part of the solution here is to eliminate the repetition of the exact same test.

As part of understanding and cleaning up this test, I also made a major effort to remove other duplication. And to improve naming and clarity. The individual test methods were lengthy, with lots of duplicated code, making it hard to understand what they were doing. Indeed some of the tests didn't actually match their name / intention. I extracted out the commonality and tried to make it clearer the variation in each test.

This test had continued testing all old versions of the JNLP protocol. There had been TODOs marked to remove them from this test. They've been deprecated for a long time. Their implementations aren't quite behaving properly, particularly when it comes to closing them down, which causes contamination with testing the currently supported protocols. With Oracle deprecating JNLP entirely it's time to remove these protocols from this test.